revert changes in .rubocop.yml file
