+++ .rubocop.yml
@@ -10,11 +10,11 @@ AllCops:
 Layout/LineLength:
   Max: 120
 Metrics/MethodLength:
-  Max: 50
+  Max: 20
 Metrics/AbcSize:
   Max: 50
 Metrics/ClassLength:
-  Max: 200
+  Max: 150
 Metrics/BlockLength:
   AllowedMethods: ["describe"]
   Max: 30
