+++ rentals.json
@@ -1,1 +1,1 @@
-{"data":[{"date":"20-20-2020","book_title":"harry potter","author":"jk rowlings","borrower_id":15},{"date":"12-12-2021","book_title":"Atomic Habit","author":"James Clear","borrower_id":374}]}
+{"data":[{"date":"20-20-2020","book_title":"harry potter","author":"jk rowlings","borrower_id":15},{"date":"12-12-2021","book_title":"Atomic Habit","author":"James Clear","borrower_id":374},{"date":"31-01-2021","book_title":"harry potter","author":"jk rowlings","borrower_id":374}]}
